+++ src/physical_ai_sandbox/policies/__init__.py
@@ -1,0 +1,16 @@
+from physical_ai_sandbox.policies.base import Policy, PolicyAction
+from physical_ai_sandbox.policies.bc import BehaviorCloningPolicy
+from physical_ai_sandbox.policies.factory import create_policy
+from physical_ai_sandbox.policies.manual import ManualPolicy
+from physical_ai_sandbox.policies.ppo import PPOPolicy
+from physical_ai_sandbox.policies.random import RandomPolicy
+
+__all__ = [
+    "BehaviorCloningPolicy",
+    "ManualPolicy",
+    "PPOPolicy",
+    "Policy",
+    "PolicyAction",
+    "RandomPolicy",
+    "create_policy",
+]